LPS-55754 Extract Journal OSGi services to Journal Service Module
