+++ modules/apps/journal/journal-service/src/com/liferay/journal/service/model/listener/GroupModelListener.java
@@ -12,28 +12,29 @@
  * details.
  */
 
-package com.liferay.portlet.journal.model;
+package com.liferay.journal.service.model.listener;
 
 import com.liferay.portal.ModelListenerException;
 import com.liferay.portal.model.BaseModelListener;
-import com.liferay.portal.model.Layout;
-import com.liferay.portlet.journal.service.JournalArticleLocalServiceUtil;
-import com.liferay.portlet.journal.service.JournalContentSearchLocalServiceUtil;
+import com.liferay.portal.model.Group;
+import com.liferay.portal.model.ModelListener;
+import com.liferay.portal.service.SubscriptionLocalServiceUtil;
+import com.liferay.portlet.journal.model.JournalArticle;
+
+import org.osgi.service.component.annotations.Component;
 
 /**
  * @author Eduardo Garcia
  */
-public class LayoutModelListener extends BaseModelListener<Layout> {
+@Component(immediate = true, service = ModelListener.class)
+public class GroupModelListener extends BaseModelListener<Group> {
 
 	@Override
-	public void onBeforeRemove(Layout layout) throws ModelListenerException {
+	public void onBeforeRemove(Group group) throws ModelListenerException {
 		try {
-			JournalArticleLocalServiceUtil.deleteLayoutArticleReferences(
-				layout.getGroupId(), layout.getUuid());
-
-			JournalContentSearchLocalServiceUtil.deleteLayoutContentSearches(
-				layout.getGroupId(), layout.isPrivateLayout(),
-				layout.getLayoutId());
+			SubscriptionLocalServiceUtil.deleteSubscriptions(
+				group.getCompanyId(), JournalArticle.class.getName(),
+				group.getGroupId());
 		}
 		catch (Exception e) {
 			throw new ModelListenerException(e);
